updating comments. getting a better grasp on my own code.
